requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0

# Graph Neural Networks
torch-geometric>=2.3.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging & Tracking
wandb>=0.15.0

# Data generation
pyyaml>=6.0

# Julia interface (for data generation)
julia>=0.6.0
